+++ app/controllers/users/registrations_controller.rb
@@ -30,10 +30,8 @@ class Users::RegistrationsController < Devise::RegistrationsController
     begin
       resource.mark_as_deleted!
 
-      # Sign out immediately
       Devise.sign_out_all_scopes ? sign_out : sign_out(resource_name)
 
-      # Enqueue background job
       Users::DestroyJob.perform_later(resource.id)
 
       set_flash_message! :notice, :destroyed